plumbing: add signature/ssh package
This package allows for the signing and verification of Git data using
SSH private and/or public keys.

It has been derived from the excellent work by the sigstore project,
which has an implementation for files (with a hardcoded `file`
namespace).

Because of this hardcoded namespace, we can not re-use the library
as-is, because the namespace used by `git` equals `git`.

Signed-off-by: Hidde Beydals <hello@hidde.co>